requirements.txt
flask==3.0.3
requests==2.32.3
faiss-cpu==1.13.2
sentence-transformers==3.0.1
unstructured[pdf]==0.14.10
numpy==1.26.4
pytest==8.3.2
